Updated the actions api to use Int64 for all sequence arguments and responses Added a contract argument to allow transfer between custom tokens

+++ eosreach/api/actions/AccountActionList.swift
@@ -2,9 +2,9 @@ import Foundation
 
 struct AccountActionList {
     let actions: [AccountAction]
-    let noResultsNext: Int
+    let noResultsNext: Int64
 
-    init(actions: [AccountAction], noResultsNext: Int = -1) {
+    init(actions: [AccountAction], noResultsNext: Int64 = -1) {
         self.actions = actions
         self.noResultsNext = noResultsNext
     }
