+++ eosreach/api/actions/AccountActionRequest.swift
@@ -5,8 +5,8 @@ protocol AccountActionsRequest {
 
     func getActionsForAccountName(
         contractAccountBalance: ContractAccountBalance,
-        position: Int,
-        offset: Int
+        position: Int64,
+        offset: Int64
     ) -> Single<Result<AccountActionList, AccountActionsError>>
 }
